+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 ## Next version
 
-- Fixed handling of scrollable text boxes in Chromium browsers.
+- Fixed handling of scrollable text boxes in Chromium and Safari browsers.
 - Made number handling a bit more tolerant so 8万8千平㍍ etc. would be recognized.
 - Fixed a case where the options page would flicker incessantly
   ([#708](https://github.com/birchill/10ten-ja-reader/discussions/708)).